2023: Day 1 - Trebuchet
This was quite a pickle for the first day! The first part was quite easy
with just a simple pattern match.

The part 2 was really giving me a hard time though! It took much longer
to figure it out than I was expecting.. eventually  I figured it out...
The solution being quite different from what this is now. Most time was
spent fine tuning the code and trying to figure out the most idiomatic
way to write it in rust.